Fix bugs and refactor tests
