+++ tests/test_github.py
@@ -1,7 +1,7 @@
 from datetime import date
 
-from ssg.github import date_only
-from ssg.stars import get_repo_state_from_handle
+from src.ssg.github import date_only
+from src.ssg.stars import get_repo_state_from_handle
 
 
 def test_get_repo_state_from_handle():
